backend/requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
sentence-transformers
torch --index-url https://download.pytorch.org/whl/cpu
numpy
